+++ shuup/admin/static_src/base/js/messages.js
@@ -8,13 +8,15 @@
  */
 window.Messages = (function Messages(document) {
     const queue = [];
-    var container = null;
+    let container = null;
+    let timeOutForHide = null;
+    let timeOutForClear = null;
+
     function createContainer() {
         if (!container) {
             container = document.createElement("div");
             container.id = "message-container";
             document.body.appendChild(container);
-            document.addEventListener("click", hideOnClickOut, false);
         }
     }
     function show() {
@@ -25,9 +27,11 @@ window.Messages = (function Messages(document) {
     }
     function hide() {
         if (container) {
-            container.classList.remove("visible");
-            container.classList.add("clear");
-            setTimeout(clear, 2000);
+          timeOutForHide = setTimeout(() => {
+              container.classList.remove("visible");
+              container.classList.add("clear");
+              timeOutForClear = setTimeout(clear, 1000);
+          }, 2000)
         }
     }
     function clear() {
@@ -36,19 +40,9 @@ window.Messages = (function Messages(document) {
             container.removeChild(container.firstChild);
         }
     }
-    function hideOnClickOut(event) {
-        var node = event.target;
-        while (node) {
-            if (node.id === "message-container") {
-                return;
-            }
-            node = node.parentNode;
-        }
-        hide();
-    }
     function renderMessage(message) {
         const messageDiv = document.createElement("div");
-        var tags = message.tags || [];
+        let tags = message.tags || [];
         if (_.isString(tags)) {
             tags = tags.split(" ");
         }
@@ -60,13 +54,12 @@ window.Messages = (function Messages(document) {
 
         const dismissIcon = document.createElement("span");
         dismissIcon.className = "dimiss-icon";
-        dismissIcon.addEventListener("click", function () {
-            messageDiv.classList.add("dimissed");
-        });
-        const statucIcon = document.createElement("span");
-        statucIcon.className = "status";
+        const statusIcon = document.createElement("span");
+        statusIcon.className = "status";
 
-        messageDiv.appendChild(statucIcon);
+        messageDiv.addEventListener("click", () => clear());
+
+        messageDiv.appendChild(statusIcon);
         messageDiv.appendChild(textSpan);
         messageDiv.appendChild(dismissIcon);
         return messageDiv;
@@ -79,10 +72,24 @@ window.Messages = (function Messages(document) {
             return setTimeout(flush, 50);
         }
         createContainer();
+
+        if (timeOutForHide !== null) {
+            // Reset the timeout for hiding the messages when a new message is added.
+            clearTimeout(timeOutForHide);
+            timeOutForHide = null;
+        }
+        if (timeOutForClear !== null) {
+            // Immediately clear the already hidden messages a when new one is added.
+            clearTimeout(timeOutForClear);
+            timeOutForClear = null;
+            clear();
+        }
+
         while (queue.length > 0) {
             container.appendChild(renderMessage(queue.shift()));
         }
         _.defer(show);
+        hide();
     }
     const deferredFlush = _.debounce(flush, 50);
     function enqueue(message) {